requirements.txt
vantage6-algorithm-tools
pandas
statsmodels
formulaic
